カラムの型合わせ

+++ database/migrations/2020_11_15_102807_change_users_table.php
@@ -14,7 +14,7 @@ class ChangeUsersTable extends Migration
     public function up()
     {
         Schema::table('users', function (Blueprint $table) {
-            $table->unsignedInteger('position_id')->after('password');
+            $table->unsignedBigInteger('position_id')->after('password');
         });
     }
